feat(#575): introduce PluginRegistrar extension point and RunServe application service
Moves plugin registration logic from cmd/vibewarden/ into internal/plugins/
and introduces internal/app/serve.RunServe as the exportable serve orchestration
entry point. The Pro binary can now register additional plugins by passing
PluginRegistrar functions to RunServe without forking the OSS codebase.
