+++ cmd/vibewarden/serve.go
@@ -2,20 +2,10 @@ package main
 
 import (
 	"context"
-	"fmt"
-	"log/slog"
-	"os"
-	"os/signal"
-	"syscall"
-	"time"
 
 	"github.com/spf13/cobra"
 
-	caddyadapter "github.com/vibewarden/vibewarden/internal/adapters/caddy"
-	logadapter "github.com/vibewarden/vibewarden/internal/adapters/log"
-	"github.com/vibewarden/vibewarden/internal/app/proxy"
-	"github.com/vibewarden/vibewarden/internal/config"
-	"github.com/vibewarden/vibewarden/internal/plugins"
+	appserve "github.com/vibewarden/vibewarden/internal/app/serve"
 )
 
 // newServeCmd creates the serve subcommand.
@@ -30,125 +20,14 @@ func newServeCmd() *cobra.Command {
 Reads configuration from vibewarden.yaml (or the path specified with --config).
 Listens for SIGINT/SIGTERM and performs a graceful shutdown.`,
 		RunE: func(cmd *cobra.Command, _ []string) error {
-			return runServe(configPath)
+			return appserve.RunServe(context.Background(), appserve.Options{
+				ConfigPath: configPath,
+				Version:    version,
+			})
 		},
 	}
 
 	cmd.Flags().StringVarP(&configPath, "config", "c", "", "path to vibewarden.yaml (default: ./vibewarden.yaml)")
 
 	return cmd
-}
-
-// runServe loads config, builds the plugin registry, wires Caddy via plugin
-// contributors, and runs until a shutdown signal is received.
-func runServe(configPath string) error {
-	cfg, err := config.Load(configPath)
-	if err != nil {
-		return fmt.Errorf("loading config: %w", err)
-	}
-
-	logger := buildLogger(cfg.Log)
-
-	// Migrate legacy metrics: config to telemetry: if needed.
-	config.MigrateLegacyMetrics(cfg, logger)
-
-	logger.Info("VibeWarden starting",
-		slog.String("version", version),
-		slog.String("listen", fmt.Sprintf("%s:%d", cfg.Server.Host, cfg.Server.Port)),
-		slog.String("upstream", fmt.Sprintf("%s:%d", cfg.Upstream.Host, cfg.Upstream.Port)),
-	)
-
-	// Build the plugin registry and register all compiled-in plugins.
-	// At registration time we use a stdout-only event logger. After InitAll,
-	// we upgrade to a multi-handler logger (stdout + OTel) when log export is
-	// enabled. The caddy adapter — the main consumer of the event logger —
-	// is created after StartAll, so it always gets the upgraded logger.
-	initialEventLogger := logadapter.NewSlogEventLogger(os.Stdout)
-
-	registry := plugins.NewRegistry(logger)
-	registerPlugins(registry, cfg, initialEventLogger, logger)
-
-	// Set up OS signal handling before Init/Start so that a slow Init can
-	// still be interrupted.
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	sigCh := make(chan os.Signal, 1)
-	signal.Notify(sigCh, syscall.SIGINT, syscall.SIGTERM)
-
-	go func() {
-		sig := <-sigCh
-		logger.Info("received shutdown signal", slog.String("signal", sig.String()))
-		cancel()
-	}()
-
-	// Initialise all plugins.
-	if err := registry.InitAll(ctx); err != nil {
-		return fmt.Errorf("initialising plugins: %w", err)
-	}
-
-	// After InitAll, retrieve the OTel log handler from the metrics plugin
-	// (if log export is enabled) and build the final event logger.
-	eventLogger := buildEventLogger(registry, logger)
-
-	// Wire the metrics collector into the TLS cert expiry monitor so that
-	// the vibewarden_tls_cert_expiry_seconds gauge is populated. This must
-	// happen after InitAll (metrics provider is ready) and before StartAll
-	// (monitor background goroutine launches on Start).
-	wireTLSMetricsCollector(registry)
-
-	// Start all plugins (background servers, etc.).
-	if err := registry.StartAll(ctx); err != nil {
-		return fmt.Errorf("starting plugins: %w", err)
-	}
-
-	// Ensure StopAll runs on return (normal or error path).
-	defer func() {
-		stopCtx, stopCancel := context.WithTimeout(context.Background(), 10*time.Second)
-		defer stopCancel()
-		if stopErr := registry.StopAll(stopCtx); stopErr != nil {
-			logger.Error("stopping plugins", slog.String("error", stopErr.Error()))
-		}
-	}()
-
-	// Build ProxyConfig — base fields that the Caddy adapter uses directly.
-	// Plugin-specific wiring (security headers, rate limiting, auth, admin,
-	// metrics) is now driven by each plugin's CaddyContributor implementation.
-	// The legacy top-level fields are still populated for the existing
-	// BuildCaddyConfig path; a follow-up issue will migrate that fully to
-	// the contributor model.
-	proxyCfg := buildProxyConfig(cfg, registry)
-
-	// Create Caddy adapter and proxy service.
-	adapter := caddyadapter.NewAdapter(proxyCfg, logger, eventLogger)
-	svc := proxy.NewService(adapter, logger)
-
-	if err := svc.Run(ctx); err != nil {
-		return fmt.Errorf("proxy service: %w", err)
-	}
-
-	return nil
-}
-
-// buildLogger creates an slog.Logger from the log configuration.
-func buildLogger(cfg config.LogConfig) *slog.Logger {
-	var level slog.Level
-	switch cfg.Level {
-	case "debug":
-		level = slog.LevelDebug
-	case "warn":
-		level = slog.LevelWarn
-	case "error":
-		level = slog.LevelError
-	default:
-		level = slog.LevelInfo
-	}
-
-	opts := &slog.HandlerOptions{Level: level}
-
-	if cfg.Format == "text" {
-		return slog.New(slog.NewTextHandler(os.Stderr, opts))
-	}
-
-	return slog.New(slog.NewJSONHandler(os.Stderr, opts))
 }
